% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_data}
\alias{all_data}
\title{Raw data from all studies}
\format{
Data frame
\describe{
\item{id_meta}{Analysis id}
}
}
\usage{
all_data
}
\description{
See data_meta.R file in data-raw folder
}
\keyword{datasets}
