% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Raw data from meta-analysis}
\format{
Data frame
\describe{
\item{id_meta}{Analysis id}
}
}
\usage{
meta
}
\description{
See data_meta.R file in data-raw folder
}
\keyword{datasets}
