% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module_meta.R
\name{metaAnalysisServer2}
\alias{metaAnalysisServer2}
\title{Server for meta-analysis module}
\usage{
metaAnalysisServer2(
  id,
  data,
  plotly_toolbar_buttons = "toImage",
  id_export_plot_btn = "export_plot_btn"
)
}
\arguments{
\item{id_export_plot_btn}{ID for button to export plot. Default is \code{"export_plot_btn"}}

\item{meta}{Data frame containing raw data and meta-analysis}
}
\value{
Enter inside server function of shiny app
}
\description{
Server for meta-analysis module
}
