% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module_immunogenicity.R
\name{metaAnalysisServer3}
\alias{metaAnalysisServer3}
\title{Server for meta-analysis module}
\usage{
metaAnalysisServer3(
  id,
  data,
  estimate_title = NULL,
  x_axis_title = "Risk ratio",
  plotly_toolbar_buttons = "toImage",
  id_export_plot_btn = "export_plot_btn",
  direction_labels = c("Favors\\nvaccination", "Favors\\ncontrol")
)
}
\arguments{
\item{id_export_plot_btn}{ID for button to export plot. Default is \code{"export_plot_btn"}}

\item{meta}{Data frame containing raw data and meta-analysis}
}
\value{
Enter inside server function of shiny app
}
\description{
Server for meta-analysis module
}
