% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module_meta.R
\name{metaAnalysisUI2}
\alias{metaAnalysisUI2}
\title{UI for meta-analysis module}
\usage{
metaAnalysisUI2(
  id,
  data,
  sidebar_open = FALSE,
  sidebar_width = "30\%",
  plot_resizable = TRUE,
  plot_width = "100\%",
  plot_height = "400px",
  show_color_picker = FALSE,
  show_legend_switch = FALSE,
  primary_color = "#246A87",
  secondary_color = "#A63B86",
  slider_color = primary_color,
  incrementor_button_color = primary_color,
  id_export_plot_btn = "export_plot_btn"
)
}
\arguments{
\item{id}{Input id. Must match \code{id} entered in \code{metaAnalysisServer2}}

\item{data}{Data frame with all analysis}

\item{sidebar_open}{If \code{TRUE} (default), sidebar is open by default}

\item{sidebar_width}{Width of sidebar. Default is \code{"30\%"}}

\item{plot_resizable}{If \code{TRUE} (default), plot can be resized by draggling bottom right corner}

\item{plot_width, plot_height}{Default width and height of plot. Default is \code{"100\%"} for \code{plot_width} and \code{"200px"} for \code{plot_height}}

\item{show_color_picker}{If \code{TRUE}, color picker is displayed above plot. Default is \code{FALSE}}

\item{show_legend_switch}{If \code{TRUE}, switch input is displayed to control whether plot legend is displayed. Default is \code{FALSE}}

\item{primary_color, secondary_color}{Primary and secondary color used for bootstrap theme. Default is \code{"#A63B86"} for \code{secondary_color}}

\item{slider_color, switch_color}{Color for sliders and switches}

\item{incrementor_button_color}{Color for incrementor button}

\item{id_export_plot_btn}{ID for export plot button. Default is \code{"export_plot_btn"}}

\item{accordion_btn_border_color}{Color for accordion buttons}

\item{accordion_btn_background_color_alpha}{Alpha filter for accordion buttons. Default is \code{0.1}}

\item{slider_color}{Color for slider and switch}
}
\value{
Enter as input to UI
}
\description{
UI for meta-analysis module
}
