% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Forest plot on linear x axis scale}
\usage{
plot_forest(
  df,
  estimate = "rr",
  lower = "rr_lower",
  upper = "rr_upper",
  y_var = NULL,
  grouping_var = NULL,
  ordering_var = NULL,
  point_size_var = NULL,
  point_size = 5,
  point_shape_var = NULL,
  point_shape = "square",
  point_color_var = grouping_var,
  point_color = "#333333",
  meta_var = "is_meta",
  meta_id = "id_meta",
  meta_label = "",
  meta_point_size = point_size * 1.05,
  meta_point_color = NULL,
  meta_point_shape = "diamond",
  errorbar_color_var = NULL,
  errorbar_color = "#333333",
  label_var = NULL,
  errorbar_width = 0.25,
  errorbar_thickness = 0.7,
  dodge_width = 0.9,
  point_border_thickness = 0.7,
  point_border_color = "#333333",
  axis_line_thickness = 0.7,
  y_axis_label_size = base_size,
  y_axis_label_color = "#333333",
  y_axis_labels = ggplot2::waiver(),
  x_axis_title = NULL,
  x_axis_limits = NULL,
  x_axis_breaks = NULL,
  x_axis_labels = ggplot2::waiver(),
  vert_line_x_position = NULL,
  vert_line_thickness = 0.5,
  vert_line_color = "#333333",
  vert_linetype = "dashed",
  legend_title = NULL,
  show_legend = FALSE,
  x_axis_scale = "regular",
  show_stripes = TRUE,
  trim_stripes = TRUE,
  odd_stripe_colors = "#22222222",
  even_stripe_colors = "#00000000",
  bracket_type = c("(", ")"),
  label_digits = 0,
  base_size = 14,
  hjust = 1,
  ratio = 2.5,
  aspect_ratio = ratio,
  expand_lower = 0,
  expand_upper = 0,
  x_axis_include = NULL,
  cap = "both",
  data_limits = NULL,
  show_x_axis = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{estimate, lower, upper}{Columns in \code{df} containing effect estimate, lower, and upper bound of confidence interval, respectively}

\item{y_var}{y axis variable. Enter as length 1 character vector. Should be a categorical variable}

\item{grouping_var}{Grouping variable along y axis}

\item{ordering_var}{Variable used to reorder rows in forest plot. If \code{NULL} (default), input is ignored}

\item{point_size_var}{Variable to determine point size}

\item{point_size}{Size of effect estimate points. Enter as length 1 numeric vector. Default is \code{5}}

\item{point_shape_var}{Variable to determine point shapes}

\item{point_shape}{Shapes used to display effect estimate. Options: \code{square} (default), \code{circle}, \code{diamond}, \code{triangle}}

\item{point_color_var}{Variable used to determine point color}

\item{point_color}{Colors for points}

\item{meta_var}{Variable to indicate whether row in \code{df} refers to meta analysis. Enter as length 1 character vector. Default is \code{"is_meta"}}

\item{meta_id}{Variable to identify rows in \code{df} included in each meta-analysis}

\item{meta_label}{Value for \code{label_var} to indicate rows with meta-analysis}

\item{meta_point_color, meta_point_shape, meta_point_size}{Color, shape, and size for meta analysis points}

\item{errorbar_color_var}{Variable to determine color of effect estimate/CI bars}

\item{errorbar_color}{Color of effect estimate/CI bars. Default is \code{"#333333"}}

\item{label_var}{Variable used to create row labels in plot}

\item{errorbar_width}{Width of error bars. Enter as length 1 numeric vector. Default is \code{0.25}}

\item{errorbar_thickness}{Line thickness for effect estimate/CI bars. Enter as length 1 numeric vector. Default is \code{0.7}}

\item{dodge_width}{Width of dodging. Only relevant when \code{grouping_var} is specified. Enter as length 1 numeric vector. Default is \code{0.9}}

\item{point_border_thickness}{Border thickness for effect estimate points. Enter as length 1 numeric vector. Default is \code{0.7}}

\item{point_border_color}{Border color (stroke) for points}

\item{axis_line_thickness}{Thickness of x axis line and ticks. Enter as length 1 numeric vector. Default is \code{0.7}}

\item{y_axis_label_size}{Font size for y axis labels. Enter as length 1 numeric vector. Default is \code{10}}

\item{y_axis_label_color}{Font color for y axis labels. Default is \code{"grey30"}}

\item{y_axis_labels}{Labels for y axis ticks}

\item{x_axis_title}{Title for x axis}

\item{x_axis_limits}{Length 2 numeric vector contains limits for x axis}

\item{x_axis_breaks, x_axis_labels}{Breaks and labels for x axis}

\item{vert_line_x_position}{x intercept of vertical reference line. Default is \code{NULL} (no dashed line)}

\item{vert_line_thickness}{Thickness of vertical reference line. Enter as length 1 numeric vector. Default is \code{0.5}}

\item{vert_line_color}{Color of vertical reference line. Enter as length 1 character vector containing hexadecimal code or named color. Default is \code{"#333333"}}

\item{vert_linetype}{Line type for vertical reference line. Default is \code{"dashed"}}

\item{legend_title}{Title for legend. Only relevant when \code{show_legend = TRUE}}

\item{show_legend}{If \code{FALSE} (default), no legend is displayed. If \code{TRUE}, legend displayed}

\item{x_axis_scale}{Scale for effect estimate. Default is \code{"regular"}}

\item{show_stripes}{If \code{TRUE} (default), alternating white and grey rectangles displayed in background. If \code{FALSE}, no stripes displayed}

\item{trim_stripes}{If \code{TRUE} (default), stripes are trimmed to axis limits. If \code{FALSE}, stripes extend to plot limits}

\item{odd_stripe_colors, even_stripe_colors}{Colors for odd and even stripes. Enter as length character vector of hexadecimal codes or colors}

\item{bracket_type}{Type of bracket to use for 95\% CI label. Default is \code{c("(", ")")}}

\item{label_digits}{Number of digits to include in label. Enter as length 1 numeric vector. Default is \code{0}}

\item{base_size}{Base font size for plot. Enter as length 1 numeric vector. Default is \code{14}}

\item{hjust}{Horizontal adjustment. Enter as length 1 numeric vector. Default is \code{1}}

\item{ratio}{Alias for \code{aspect_ratio}}

\item{aspect_ratio}{Aspect ratio of plot. Enter as length 1 numeric vector. Default is \code{2.5}}

\item{expand_lower, expand_upper}{Expansion for lower and upper end of x axis. Enter each as length 1 numeric vector}

\item{x_axis_include}{Value to include along x axis. Enter as numeric vector. Only relevant when \code{x_axis_limits = NULL}}

\item{cap}{Cap for x axis. Default is \code{"both"}}

\item{data_limits}{x axis limits for error bars. Only changes plot appearance, not actually estimates/labels. Enter as length 2 numeric vector. Default is \code{NULL}, no limits applied}

\item{show_x_axis}{If \code{TRUE} (default), x axis is displayed. If \code{FALSE}, no scale is applied to x axis}

\item{...}{Arguments passed to \code{theme_vip}}

\item{ci_sep}{Separator between lower and upper bound in 95\% CI labels. Enter as length 1 character vector. Default is \code{"-"}. Forest plots for manuscript used \code{" to "}}
}
\value{
ggplot object
}
\description{
Forest plot on linear x axis scale
}
